intente hacer andar el browser routing pero me tira error en el navLink por el boostrap pero no logro entender porque no anda
