+++ package.json
@@ -10,6 +10,7 @@
     "react": "^18.2.0",
     "react-bootstrap": "^2.4.0",
     "react-dom": "^18.2.0",
+    "react-router-bootstrap": "^0.26.2",
     "react-scripts": "5.0.1",
     "web-vitals": "^2.1.4"
   },
