+++ src/components/ItemListContainer.js
@@ -1,5 +1,4 @@
 import { useEffect, useState } from "react"
-import ItemDetail from "./ItemDetail"
 import ItemList from "./ItemList"
 
 const productosIniciales = [
@@ -53,7 +52,6 @@ const ItemListContainer = () => {
     } else {
         return (
             <ItemList productos={productos}>
-                <ItemDetail detail = {detail}></ItemDetail>
             </ItemList>
         )
     }
