+++ src/components/Main.js
@@ -1,12 +1,15 @@
 import React from 'react'
 import ItemListContainer from "./ItemListContainer"
+import { Routes , Route } from "react-router-dom"
 import ItemDetailContainer from "./ItemDetailContainer"
-
 
 const Main = () => {
   return (
     <div>
-        <ItemListContainer/>
+            <Routes>
+                <Route path="/" element={<ItemListContainer/>}/>
+                <Route path="/detalle" element={<ItemDetailContainer/>}/>
+            </Routes>
     </div>
   )
 }
